GetNevents to long int instead of int (maximum was too small)

+++ Utilities/AnalysisUtilities.C
@@ -36,11 +36,11 @@ void TransformRawHistToYield(TH1D* histogram){
   histogram->Scale(1.,"width"); // If option contains "width" the bin contents and errors are divided by the bin width.
 }
 
-int GetNEventsSel8(TFile* file_O2Analysis) {
+long int GetNEventsSel8(TFile* file_O2Analysis) {
   return ((TH1I*)file_O2Analysis->Get("event-selection-task/hColCounterAcc"))->GetEntries(); //this is only sel8 (no sel8Full for example) and doesn't exclude collisions cut by the vertexZ pos cut
 }
 
-int GetNEventsSelected_JetFramework(TFile* file_O2Analysis, const char analysisWorkflow[]) {
+long int GetNEventsSelected_JetFramework(TFile* file_O2Analysis, const char analysisWorkflow[]) {
   return ((TH1I*)file_O2Analysis->Get((TString)analysisWorkflow+"/h_collisions"))->GetBinContent(3); //this should be the actual selection AND take vertexZ into account;
 }
 double GetNEventsSelected_JetFramework_weighted(TFile* file_O2Analysis, const char analysisWorkflow[]) {
@@ -48,24 +48,24 @@ double GetNEventsSelected_JetFramework_weighted(TFile* file_O2Analysis, const ch
 }
 
 
-int GetNEventsSelected_TrackEffWorkflow(TFile* file_O2Analysis, const char analysisWorkflow[]) {
+long int GetNEventsSelected_TrackEffWorkflow(TFile* file_O2Analysis, const char analysisWorkflow[]) {
   return ((TH1I*)file_O2Analysis->Get((TString)analysisWorkflow+"/h_collisions"))->GetBinContent(3); //this should be the actual selection AND take vertexZ into account; 
 }
 double GetNEventsSelected_TrackEffWorkflow_weighted(TFile* file_O2Analysis, const char analysisWorkflow[]) {
   return ((TH1F*)file_O2Analysis->Get((TString)analysisWorkflow+"/h_collisions_weighted"))->GetBinContent(3); //this should be the actual selection AND take vertexZ into account; 
 }
-int GetNEventsSelected_TrackEffWorkflow_gen(TFile* file_O2Analysis, const char analysisWorkflow[]) {
+long int GetNEventsSelected_TrackEffWorkflow_gen(TFile* file_O2Analysis, const char analysisWorkflow[]) {
   return ((TH1I*)file_O2Analysis->Get((TString)analysisWorkflow+"/h_mccollisions"))->GetBinContent(2); //this should be the actual selection AND take vertexZ into account;
 }
 double GetNEventsSelected_TrackEffWorkflow_gen_weighted(TFile* file_O2Analysis, const char analysisWorkflow[]) {
   return ((TH1F*)file_O2Analysis->Get((TString)analysisWorkflow+"/h_mccollisions_weighted"))->GetBinContent(2); //this should be the actual selection AND take vertexZ into account;
 }
 
-int GetNEventsGen(TFile* file_O2Analysis) {
+long int GetNEventsGen(TFile* file_O2Analysis) {
   return ((TH1I*)file_O2Analysis->Get("event-selection-task/hColCounterAcc"))->GetEntries();
 }
 
-int GetNEventsSel8Centrality(TFile* file_O2Analysis, float centralityLow, float centralityHigh) { // probably same issue as getneventssel8
+long int GetNEventsSel8Centrality(TFile* file_O2Analysis, float centralityLow, float centralityHigh) { // probably same issue as getneventssel8
   TH1D* H1D_Centrality_FT0C= (TH1D*)file_O2Analysis->Get("centrality-qa/hCentFT0C");
   int iBinCent_low = H1D_Centrality_FT0C->GetXaxis()->FindBin(centralityLow + GLOBAL_epsilon);
   int iBinCent_high = H1D_Centrality_FT0C->GetXaxis()->FindBin(centralityHigh - GLOBAL_epsilon);
@@ -73,7 +73,7 @@ int GetNEventsSel8Centrality(TFile* file_O2Analysis, float centralityLow, float 
 }
 
 
-int GetNEventsSelectedCentrality_JetFramework(TFile* file_O2Analysis, const char analysisWorkflow[], float centralityLow, float centralityHigh) { // should check it gives the correct number of coll (is posZ taken into account, fullsel8 etc)
+long int GetNEventsSelectedCentrality_JetFramework(TFile* file_O2Analysis, const char analysisWorkflow[], float centralityLow, float centralityHigh) { // should check it gives the correct number of coll (is posZ taken into account, fullsel8 etc)
   TH2D* H2D_Centrality_FT0C= (TH2D*)file_O2Analysis->Get((TString)analysisWorkflow+"/h2_centrality_collisions");
   TH1D* H1D_Centrality_FT0C= (TH1D*)H2D_Centrality_FT0C->ProjectionX("H1D_Centrality_FT0C", 2, 2);
